+++ Server/demo/src/main/java/com/example/demo/entity/Mark.java
@@ -11,38 +11,38 @@ import java.io.Serializable;
 public class Mark extends Model<Mark> {
     private static final long serialVersionUID = 1L;
 
-    private Integer id;
-    private Integer uid;
-    private Integer hid;
+    private Long id;
+    private Long uid;
+    private Long hid;
 
     @Override
     protected Serializable pkVal() {
         return this.id;
     }
 
-    public Integer getHid() {
+    public Long getHid() {
         return hid;
     }
 
-    public Mark setHid(Integer hid) {
+    public Mark setHid(Long hid) {
         this.hid = hid;
         return this;
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public Mark setId(Integer id) {
+    public Mark setId(Long id) {
         this.id = id;
         return this;
     }
 
-    public Integer getUid() {
+    public Long getUid() {
         return uid;
     }
 
-    public Mark setUid(Integer uid) {
+    public Mark setUid(Long uid) {
         this.uid = uid;
         return this;
     }
