+++ Server/demo/src/main/java/com/example/demo/entity/Users.java
@@ -19,7 +19,7 @@ public class Users extends Model<Users> {
     private static final long serialVersionUID = 1L;
 
 	@TableId(value="id", type= IdType.AUTO)
-	private Integer id;
+	private Long id;
 	private String name;
 	@TableField("phone_number")
 	private String phoneNumber;
@@ -34,11 +34,11 @@ public class Users extends Model<Users> {
 
 	}
 
-	public Integer getId() {
+	public Long getId() {
 		return id;
 	}
 
-	public Users setId(Integer id) {
+	public Users setId(Long id) {
 		this.id = id;
 		return this;
 	}
